fix: sequence cache and last value stays within minvalue and maxvalue
wip
This commit addresss an issue when increasing the sequence value. If the
next value increase beyond the minvalue or maxvalue, the cache size is
decreased to keep the sequence value within the range.
To ensure that we do not go beyond the sequence limit, the sequence
value is fetched first, calculations are done and then cput is used to
write the new sequence value. Since it is possible for this process to
be executed concurrently, the writing of the value is in a retry loop
which will not fail if the error return is from cput. This process can
be considered to be similair to a transaction as it ensures consistency.

Release note (bug fix): Sequence fetchValues ensures that the the next
value does not go beyond sequence limits.